Moves Product items to product component folder, reflects changes in router
